refactor: move components to another folders
