--- src/components/__tests__/HelloWorld.spec.js
@@ -1,11 +1,0 @@
-import { describe, it, expect } from "vitest";
-
-import { mount } from "@vue/test-utils";
-import HelloWorld from "../HelloWorld.vue";
-
-describe("HelloWorld", () => {
-  it("renders properly", () => {
-    const wrapper = mount(HelloWorld, { props: { msg: "Hello Vitest" } });
-    expect(wrapper.text()).toContain("Hello Vitest");
-  });
-});
