+++ src/components/icons/icon-app-logo.vue
