+++ src/components/metrics/metrics-picture-name-card.vue
