+++ src/components/shared/header/header-browse.vue
@@ -11,7 +11,7 @@
       <icon-menu class="header__menu-icon" />
     </button>
     <router-link class="header__link" :to="{ name: 'browse' }">
-      <app-logo-icon class="header__logo" />
+      <icon-app-logo class="header__logo" />
     </router-link>
     <appear-from from="right">
       <movies-lateral-menu v-if="menuOpen" class="header__menu" />
@@ -28,7 +28,7 @@
       class="header__link header__link--match885px"
       :to="{ name: 'browse' }"
     >
-      <app-logo-icon class="header__logo header__logo--match885px" />
+      <icon-app-logo class="header__logo header__logo--match885px" />
     </router-link>
     <nav class="header__nav">
       <ul class="header__list">
@@ -56,10 +56,10 @@
   import { useBlobURL } from "@/composables/useBlobURL";
   import { useMatchMedia } from "@/composables/useMatchMedia";
   import { ref, onMounted, onUnmounted } from "vue";
-  import { useLoginStore } from "../../stores/login";
-  import AppLogoIcon from "../icons/app-logo-icon.vue";
-  import IconMenu from "../icons/icon-menu.vue";
-  import AppearFrom from "../shared/transitions/appear-from.vue";
+  import { useLoginStore } from "@/stores/login";
+  import IconAppLogo from "@/components/icons/icon-app-logo.vue";
+  import IconMenu from "@/components/icons/icon-menu.vue";
+  import AppearFrom from "@/components/shared/transitions/appear-from.vue";
   import MoviesLateralMenu from "./header-lateral-menu.vue";
   const loginStore = useLoginStore();
 
@@ -83,7 +83,7 @@
 </script>
 
 <style lang="scss" scoped>
-  @use "../../styles/colors.scss" as *;
+  @use "@/styles/colors.scss" as *;
   .header {
     z-index: 2;
     width: 100vw;
