+++ src/components/shared/movies/movie-card.vue
