+++ src/components/shared/movies/movie-main.vue
