+++ src/components/shared/movies/movies-main-section.vue
@@ -10,7 +10,7 @@
 
 <script setup lang="ts">
   import type { Movie as MovieProp } from "@/database/database";
-  import ListHorizontal from "../utils/list-horizontal.vue";
+  import ListHorizontal from "@/components/shared/list/list-horizontal.vue";
   import MovieCard from "./movie-card.vue";
 
   interface Props {
