+++ src/views/profile-view.vue
@@ -30,13 +30,13 @@
 </template>
 
 <script setup lang="ts">
-  import HeaderBrowse from "../components/header/header-browse.vue";
+  import HeaderBrowse from "@/components/shared/header/header-browse.vue";
   import { useLoginStore } from "../stores/login";
-  import TitleTextLabel from "../components/utils/title-text-label.vue";
+  import TitleTextLabel from "@/components/profile/profile-text-field.vue";
   import { onMounted, ref } from "vue";
   import { useBlobURL } from "@/composables/useBlobURL";
   import { db, type Movie } from "@/database/database";
-  import MovieCard from "@/components/movies/movie-card.vue";
+  import MovieCard from "@/components/shared/movies/movie-card.vue";
 
   const loginStore = useLoginStore();
   const url = useBlobURL(loginStore.picture!);
